+++ lib/omni_auth/strategies/sso_oauth2_provider.rb
@@ -28,7 +28,7 @@ module OmniAuth::Strategies
 
     def authorize_params
       super.tap do |params|
-        params[:scope] = ENV['OAUTH2_SCOPE']
+        params[:scope] = active_provider.providable.scopes
       end
     end